+++ code/game/g_active.c
@@ -1169,12 +1169,17 @@ void BaseJK2_ClientThink_real( gentity_t *ent ) { // Tr!Force: [BaseJK2] Client 
 			duelAgainst->client->ps.duelIndex != ent->s.number)
 		{
 			ent->client->ps.duelInProgress = 0;
+			ent->client->pers.JKPlusForceDuel = 0; // Tr!Force: [Duel] Turn off force duels
 			G_AddEvent(ent, EV_PRIVATE_DUEL, 0);
 		}
 		else if (duelAgainst->health < 1 || duelAgainst->client->ps.stats[STAT_HEALTH] < 1)
 		{
 			ent->client->ps.duelInProgress = 0;
 			duelAgainst->client->ps.duelInProgress = 0;
+
+			// Tr!Force: [Duel] Turn off force duels
+			ent->client->pers.JKPlusForceDuel = 0;
+			duelAgainst->client->pers.JKPlusForceDuel = 0;
 
 			G_AddEvent(ent, EV_PRIVATE_DUEL, 0);
 			G_AddEvent(duelAgainst, EV_PRIVATE_DUEL, 0);
@@ -1236,6 +1241,10 @@ void BaseJK2_ClientThink_real( gentity_t *ent ) { // Tr!Force: [BaseJK2] Client 
 			{
 				ent->client->ps.duelInProgress = 0;
 				duelAgainst->client->ps.duelInProgress = 0;
+
+				// Tr!Force: [Duel] Turn off force duels
+				ent->client->pers.JKPlusForceDuel = 0;
+				duelAgainst->client->pers.JKPlusForceDuel = 0;
 
 				G_AddEvent(ent, EV_PRIVATE_DUEL, 0);
 				G_AddEvent(duelAgainst, EV_PRIVATE_DUEL, 0);